move storage types and related utilities to a separate header
